major updates to port and Port
