+++ interfaces.go
@@ -5,7 +5,7 @@ import (
 	"time"
 )
 
-type Port interface {
+type port interface {
 	io.ReadWriteCloser
 	Inwaiting() int
 	SetDeadline(time.Time) error
