+++ port_linux.go
@@ -1,27 +1,36 @@
 package serial
 
-import "os"
+import (
+	"os"
+	"time"
+)
 
-type port struct {
-	f os.File
+type Port struct {
+	f *os.File
 }
 
-func (p *port) Read(b []byte) (int, error) {
+func (p *Port) Read(b []byte) (int, error) {
 	return p.f.Read(b)
 }
 
-func (p *port) Write(b []byte) error {
+func (p *Port) Write(b []byte) (int, error) {
 	return p.f.Write(b)
 }
 
-func (p *port) Close() error {
+func (p *Port) Close() error {
 	return p.f.Close()
 }
 
-func (p *port) InWaiting() int {
+func (p *Port) InWaiting() int {
 	// Funky time
+	return 0
 }
 
-funct (p *port) SetDeadline(t time.Time) error {
+func (p *Port) SetDeadline(t time.Time) error {
 	// Funky Town
+	return nil
+}
+
+func NewPort(f *os.File) *Port {
+	return &Port{f}
 }
